mk-mmc:am335x: Prepare sdcard with 4 partitions
Jelly Bean 4.2 requires emulated storage to be ext4
for security reasons. So this partition cannot be used
to preload media content as was done in previous versions.

Modify the mkmmc-android script to create a 4th partition
in sdcard for storing preloaded media files. This partition
is mounted as secondary storage in the android system.

The emulated sdcard partition (ext4) is renamed to usrdata
and the new FAT partition is named data for continuity with
previous releases.

The new sdcard partitioning scheme is as follows:

1. boot    : FAT (MLO, u-boot, kernel)
2. rootfs  : ext4
3. usrdata : ext4 (Android internal storage, emulated sdcard)
4. data    : FAT

Change-Id: I470e593df7273a5f42fed73badbd3d525d7a7d1c
Signed-off-by: Vishveshwar Bhat <vishveshwar.bhat@ti.com>